fix: moved promt to summarizer
Prompt was not a part of summarizer package and because of that there was an error "No module named "prompt"

Moved it to summarizer directory which temporary should fix the issue.

Ultimately, I need to think about how the structure of these files and modules should look like.
