+++ Projects/Nitrodigest/Cli/setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = nitrodigest-cli
-version = 0.1.1
+version = 0.1.2
 author = Marcin Kwiatkowski
 author_email = marcin@frodigo.com
 description = The privacy‑first, local‑LLM text‑summariser for developers.
