+++ Projects/Nitrodigest/Cli/src/__init__.py
@@ -1,2 +1,2 @@
 """nitrodigest CLI package"""
-__version__ = "0.1.1"
+__version__ = "0.1.2"
