+++ Projects/Nitrodigest/Cli/src/summarizer/prompt_template.txt
